create routes folder and move files
